test: split oversized webhook tests
Generated-By: looper 0.0.0-dev (runner=fixer, agent=opencode)
