+++ internal/cliapp/webhook_commands_test.go
@@ -7,13 +7,10 @@ import (
 	"net/http"
 	"net/http/httptest"
 	"os"
-	"path/filepath"
 	"strings"
 	"testing"
 	"time"
 
-	"github.com/nexu-io/looper/internal/config"
-	"github.com/nexu-io/looper/internal/storage"
 	pkgapi "github.com/nexu-io/looper/pkg/api"
 )
 
@@ -299,168 +296,6 @@ func TestWebhookStatusRestartRequiredFalseWhenConfigMatchesRuntime(t *testing.T)
 		t.Fatalf("Run(webhook status --json) exit code = %d, want 0; stderr=%q", exitCode, stderr)
 	}
 	assertJSONContains(t, stdout, "restartRequired", false)
-}
-
-func TestWebhookStatusRestartRequiredTracksTunnelEndpointDrift(t *testing.T) {
-	t.Parallel()
-
-	server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		if r.URL.Path != "/api/v1/webhook/status" {
-			t.Fatalf("request path = %q, want %q", r.URL.Path, "/api/v1/webhook/status")
-		}
-		writeEnvelope(t, w, pkgapi.Success("req_webhook", map[string]any{
-			"enabled":                     true,
-			"mode":                        "tunnel",
-			"listenerPath":                "/webhook/forward",
-			"endpointUrl":                 "http://127.0.0.1:17310/webhook/forward",
-			"tunnelListenerUrl":           "http://127.0.0.1:8443",
-			"tunnelPublicBaseUrl":         "https://runtime.example.com/base",
-			"fallbackPollIntervalSeconds": 300,
-			"degraded":                    false,
-			"degradedReasons":             []string{},
-			"queue":                       map[string]any{"pending": 0, "capacity": 8, "activeWorkers": 0},
-			"counters":                    map[string]any{"deliveriesReceived": 0, "coalesced": 0, "dropped": 0, "queued": 0, "processed": 0, "failed": 0},
-			"recentOutcomes":              []map[string]any{},
-			"forwarders":                  []map[string]any{},
-			"tunnelHooks":                 []map[string]any{},
-		}))
-	}))
-	defer server.Close()
-
-	configPath := writeEditableCLIConfigWithPayload(t, map[string]any{
-		"webhook": map[string]any{"enabled": true, "mode": "tunnel", "listenPort": 9443, "publicBaseUrl": "https://config.example.com/base", "fallbackPollIntervalSeconds": 300},
-		"server":  map[string]any{"baseUrl": server.URL, "authMode": "none"},
-		"notifications": map[string]any{
-			"osascript": map[string]any{"enabled": false},
-		},
-	})
-	exitCode, stdout, stderr := runApp(t, "webhook", "status", "--json", "--config", configPath)
-	if exitCode != 0 {
-		t.Fatalf("Run(webhook status --json) exit code = %d, want 0; stderr=%q", exitCode, stderr)
-	}
-	assertJSONContains(t, stdout, "restartRequired", true)
-}
-
-func TestWebhookStatusRestartRequiredTracksTunnelEndpointDriftForMixedModeProjects(t *testing.T) {
-	t.Parallel()
-
-	server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		if r.URL.Path != "/api/v1/webhook/status" {
-			t.Fatalf("request path = %q, want %q", r.URL.Path, "/api/v1/webhook/status")
-		}
-		writeEnvelope(t, w, pkgapi.Success("req_webhook", map[string]any{
-			"enabled":                     true,
-			"mode":                        "gh-forward",
-			"listenerPath":                "/webhook/forward",
-			"endpointUrl":                 "http://127.0.0.1:17310/webhook/forward",
-			"tunnelListenerUrl":           "http://127.0.0.1:8443",
-			"tunnelPublicBaseUrl":         "https://runtime.example.com/base",
-			"fallbackPollIntervalSeconds": 300,
-			"degraded":                    false,
-			"degradedReasons":             []string{},
-			"queue":                       map[string]any{"pending": 0, "capacity": 8, "activeWorkers": 0},
-			"counters":                    map[string]any{"deliveriesReceived": 0, "coalesced": 0, "dropped": 0, "queued": 0, "processed": 0, "failed": 0},
-			"recentOutcomes":              []map[string]any{},
-			"forwarders":                  []map[string]any{},
-			"tunnelHooks":                 []map[string]any{},
-		}))
-	}))
-	defer server.Close()
-
-	configPath := writeEditableCLIConfigWithPayload(t, map[string]any{
-		"webhook":  map[string]any{"enabled": true, "mode": "gh-forward", "listenPort": 9443, "publicBaseUrl": "https://config.example.com/base", "fallbackPollIntervalSeconds": 300},
-		"projects": []map[string]any{{"id": "proj-1", "name": "Looper", "repoPath": t.TempDir(), "webhook": map[string]any{"mode": "tunnel"}}},
-		"server":   map[string]any{"baseUrl": server.URL, "authMode": "none"},
-		"notifications": map[string]any{
-			"osascript": map[string]any{"enabled": false},
-		},
-	})
-	exitCode, stdout, stderr := runApp(t, "webhook", "status", "--json", "--config", configPath)
-	if exitCode != 0 {
-		t.Fatalf("Run(webhook status --json) exit code = %d, want 0; stderr=%q", exitCode, stderr)
-	}
-	assertJSONContains(t, stdout, "restartRequired", true)
-}
-
-func TestWebhookStatusRestartRequiredTracksTunnelProjectSetDrift(t *testing.T) {
-	t.Parallel()
-
-	server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		if r.URL.Path != "/api/v1/webhook/status" {
-			t.Fatalf("request path = %q, want %q", r.URL.Path, "/api/v1/webhook/status")
-		}
-		writeEnvelope(t, w, pkgapi.Success("req_webhook", map[string]any{
-			"enabled":                     true,
-			"mode":                        "gh-forward",
-			"configuredTunnelProjectIds":  []string{},
-			"listenerPath":                "/webhook/forward",
-			"endpointUrl":                 "http://127.0.0.1:17310/webhook/forward",
-			"tunnelListenerUrl":           "http://127.0.0.1:9443",
-			"tunnelPublicBaseUrl":         "https://config.example.com/base",
-			"fallbackPollIntervalSeconds": 300,
-			"degraded":                    false,
-			"degradedReasons":             []string{},
-			"queue":                       map[string]any{"pending": 0, "capacity": 8, "activeWorkers": 0},
-			"counters":                    map[string]any{"deliveriesReceived": 0, "coalesced": 0, "dropped": 0, "queued": 0, "processed": 0, "failed": 0},
-			"recentOutcomes":              []map[string]any{},
-			"forwarders":                  []map[string]any{},
-			"tunnelHooks":                 []map[string]any{},
-		}))
-	}))
-	defer server.Close()
-
-	configPath := writeEditableCLIConfigWithPayload(t, map[string]any{
-		"webhook":  map[string]any{"enabled": true, "mode": "gh-forward", "listenPort": 9443, "publicBaseUrl": "https://config.example.com/base", "fallbackPollIntervalSeconds": 300},
-		"projects": []map[string]any{{"id": "proj-1", "name": "Looper", "repoPath": t.TempDir(), "webhook": map[string]any{"mode": "tunnel"}}},
-		"server":   map[string]any{"baseUrl": server.URL, "authMode": "none"},
-		"notifications": map[string]any{
-			"osascript": map[string]any{"enabled": false},
-		},
-	})
-	exitCode, stdout, stderr := runApp(t, "webhook", "status", "--json", "--config", configPath)
-	if exitCode != 0 {
-		t.Fatalf("Run(webhook status --json) exit code = %d, want 0; stderr=%q", exitCode, stderr)
-	}
-	assertJSONContains(t, stdout, "restartRequired", true)
-}
-
-func TestWebhookWarningsIncludesLoopbackWarningForMixedModeGHForwardOverrides(t *testing.T) {
-	t.Parallel()
-
-	warnings := webhookWarnings(config.Config{
-		Server:  config.ServerConfig{Host: "0.0.0.0"},
-		Webhook: config.WebhookConfig{Mode: config.WebhookModeTunnel},
-		Projects: []config.ProjectRefConfig{{
-			ID:       "proj-1",
-			Name:     "Looper",
-			RepoPath: t.TempDir(),
-			Webhook:  config.ProjectWebhookConfig{Mode: config.WebhookModeGHForward},
-		}},
-	})
-
-	if len(warnings) == 0 || warnings[0] != "server.host is not loopback; looperd will degrade webhook mode to poll fallback" {
-		t.Fatalf("webhookWarnings() = %v, want loopback warning for mixed-mode gh-forward overrides", warnings)
-	}
-}
-
-func TestWebhookWarningsSkipsLoopbackWarningForPureTunnelConfigs(t *testing.T) {
-	t.Parallel()
-
-	warnings := webhookWarnings(config.Config{
-		Server:  config.ServerConfig{Host: "0.0.0.0"},
-		Webhook: config.WebhookConfig{Mode: config.WebhookModeTunnel},
-		Projects: []config.ProjectRefConfig{{
-			ID:       "proj-1",
-			Name:     "Looper",
-			RepoPath: t.TempDir(),
-		}},
-	})
-
-	for _, warning := range warnings {
-		if warning == "server.host is not loopback; looperd will degrade webhook mode to poll fallback" {
-			t.Fatalf("webhookWarnings() = %v, want no loopback warning when every project uses tunnel", warnings)
-		}
-	}
 }
 
 func TestWebhookStatusVerboseShowsRuntimeDetails(t *testing.T) {
@@ -773,175 +608,4 @@ func TestWebhookCleanupConfirmContinuesPastMissingShownHook(t *testing.T) {
 	if !strings.Contains(stdout.String(), "Deleted 2 GitHub CLI webhook hook(s) for acme/looper.") {
 		t.Fatalf("stdout = %q, want delete confirmation after continuing past a missing hook", stdout.String())
 	}
-}
-
-func TestWebhookRotateRefusesURLMismatchBeforePatch(t *testing.T) {
-	t.Parallel()
-
-	configPath, repo := writeWebhookCommandConfigWithRecord(t, storage.WebhookTunnelHookRecord{Repo: "acme/looper", HookID: 42, ManagedURL: "https://example.com/webhook/acme/looper", SecretRef: "webhook_acme_looper.key", CreatedAt: 1, UpdatedAt: 1})
-	stdout := &bytes.Buffer{}
-	stderr := &bytes.Buffer{}
-	commands := []string{}
-	app := New(Deps{
-		Stdout: stdout,
-		Stderr: stderr,
-		LookPath: func(command string) (string, error) {
-			if command == "gh" {
-				return "/usr/bin/gh", nil
-			}
-			return command, nil
-		},
-		RunCommand: func(ctx context.Context, command string, args []string, timeout time.Duration) (commandExecutionResult, error) {
-			commands = append(commands, command+" "+strings.Join(args, " "))
-			if len(commands) != 1 {
-				t.Fatalf("unexpected RunCommand call %d: %s %q", len(commands), command, args)
-			}
-			return commandExecutionResult{Stdout: `{"id":42,"active":true,"events":["pull_request"],"config":{"url":"https://example.com/other","content_type":"json","insecure_ssl":"0"}}`, ExitCode: 0}, nil
-		},
-	})
-
-	exitCode := app.Run(context.Background(), []string{"webhook", "rotate", repo, "--config", configPath})
-	if exitCode == 0 {
-		t.Fatalf("Run(webhook rotate) exit code = %d, want non-zero", exitCode)
-	}
-	if len(commands) != 1 || !strings.HasSuffix(commands[0], " api repos/acme/looper/hooks/42") {
-		t.Fatalf("commands = %q, want a single GET preflight", commands)
-	}
-	if !strings.Contains(stderr.String(), "refusing to rotate hook 42 for acme/looper") {
-		t.Fatalf("stderr = %q, want URL mismatch refusal", stderr.String())
-	}
-}
-
-func TestWebhookRotateFailsWhenExistingSecretCannotBeRead(t *testing.T) {
-	t.Parallel()
-
-	configPath, repo := writeWebhookCommandConfigWithRecord(t, storage.WebhookTunnelHookRecord{Repo: "acme/looper", HookID: 42, ManagedURL: "https://example.com/webhook/acme/looper", SecretRef: "webhook_acme_looper.key", CreatedAt: 1, UpdatedAt: 1})
-	stdout := &bytes.Buffer{}
-	stderr := &bytes.Buffer{}
-	commands := []string{}
-	app := New(Deps{
-		Stdout: stdout,
-		Stderr: stderr,
-		LookPath: func(command string) (string, error) {
-			if command == "gh" {
-				return "/usr/bin/gh", nil
-			}
-			return command, nil
-		},
-		RunCommand: func(ctx context.Context, command string, args []string, timeout time.Duration) (commandExecutionResult, error) {
-			commands = append(commands, command+" "+strings.Join(args, " "))
-			if len(commands) != 1 {
-				t.Fatalf("unexpected RunCommand call %d: %s %q", len(commands), command, args)
-			}
-			return commandExecutionResult{Stdout: `{"id":42,"active":true,"events":["pull_request"],"config":{"url":"https://example.com/webhook/acme/looper","content_type":"json","insecure_ssl":"0"}}`, ExitCode: 0}, nil
-		},
-	})
-
-	secretPath := webhookTunnelSecretPathForCLI(filepath.Join(filepath.Dir(configPath), "looper.sqlite"), "webhook_acme_looper.key")
-	if _, err := os.Stat(secretPath); !os.IsNotExist(err) {
-		t.Fatalf("os.Stat(%q) error = %v, want not exists", secretPath, err)
-	}
-
-	exitCode := app.Run(context.Background(), []string{"webhook", "rotate", repo, "--config", configPath})
-	if exitCode == 0 {
-		t.Fatalf("Run(webhook rotate) exit code = %d, want non-zero", exitCode)
-	}
-	if len(commands) != 1 || !strings.HasSuffix(commands[0], " api repos/acme/looper/hooks/42") {
-		t.Fatalf("commands = %q, want a single GET preflight before the local secret read failure", commands)
-	}
-	if !strings.Contains(stderr.String(), "read existing tunnel webhook secret") {
-		t.Fatalf("stderr = %q, want secret read failure", stderr.String())
-	}
-	if _, err := os.Stat(secretPath); !os.IsNotExist(err) {
-		t.Fatalf("os.Stat(%q) error = %v, want secret file to remain missing", secretPath, err)
-	}
-	if stdout.Len() != 0 {
-		t.Fatalf("stdout = %q, want empty on failure", stdout.String())
-	}
-}
-
-func TestWebhookDeleteConfirmForgetRemovesLocalRecordWithoutRunningGH(t *testing.T) {
-	t.Parallel()
-
-	configPath, repoStoreKey := writeWebhookCommandConfigWithRecord(t, storage.WebhookTunnelHookRecord{Repo: "acme/looper", HookID: 42, ManagedURL: "https://example.com/webhook/acme/looper", SecretRef: "webhook_acme_looper.key", Orphaned: true, CreatedAt: 1, UpdatedAt: 1})
-	stdout := &bytes.Buffer{}
-	stderr := &bytes.Buffer{}
-	runCalled := false
-	app := New(Deps{
-		Stdout: stdout,
-		Stderr: stderr,
-		LookPath: func(command string) (string, error) {
-			if command == "gh" {
-				return "/usr/bin/gh", nil
-			}
-			return command, nil
-		},
-		RunCommand: func(ctx context.Context, command string, args []string, timeout time.Duration) (commandExecutionResult, error) {
-			runCalled = true
-			return commandExecutionResult{}, nil
-		},
-	})
-
-	exitCode := app.Run(context.Background(), []string{"webhook", "delete", repoStoreKey, "--confirm", "--forget", "--config", configPath})
-	if exitCode != 0 {
-		t.Fatalf("Run(webhook delete --confirm --forget) exit code = %d, want 0; stderr=%q", exitCode, stderr.String())
-	}
-	if runCalled {
-		t.Fatal("RunCommand() called, want no gh invocation")
-	}
-	if !strings.Contains(stdout.String(), "Forgot local tunnel webhook record for acme/looper") {
-		t.Fatalf("stdout = %q, want forget confirmation", stdout.String())
-	}
-	db, err := storage.OpenSQLiteDB(context.Background(), filepath.Join(filepath.Dir(configPath), "looper.sqlite"))
-	if err != nil {
-		t.Fatalf("OpenSQLiteDB() error = %v", err)
-	}
-	defer func() { _ = db.Close() }()
-	_, ok, err := storage.NewRepositories(db).WebhookTunnelHooks.Get(context.Background(), "acme/looper")
-	if err != nil {
-		t.Fatalf("WebhookTunnelHooks.Get() error = %v", err)
-	}
-	if ok {
-		t.Fatal("WebhookTunnelHooks.Get() found record, want deleted")
-	}
-}
-
-func writeWebhookCommandConfigWithRecord(t *testing.T, record storage.WebhookTunnelHookRecord) (string, string) {
-	t.Helper()
-
-	dir := t.TempDir()
-	configPath := filepath.Join(dir, "config.json")
-	dbPath := filepath.Join(dir, "looper.sqlite")
-	coordinator := openMigratedCLIWebhookCoordinator(t, dbPath)
-	defer func() { _ = coordinator.Close() }()
-	if err := storage.NewRepositories(coordinator.DB()).WebhookTunnelHooks.Upsert(context.Background(), record); err != nil {
-		t.Fatalf("WebhookTunnelHooks.Upsert() error = %v", err)
-	}
-	raw, err := json.Marshal(map[string]any{
-		"storage": map[string]any{"dbPath": dbPath},
-		"tools":   map[string]any{"ghPath": "/usr/bin/gh"},
-		"notifications": map[string]any{
-			"osascript": map[string]any{"enabled": false},
-		},
-	})
-	if err != nil {
-		t.Fatalf("marshal config: %v", err)
-	}
-	if err := os.WriteFile(configPath, raw, 0o644); err != nil {
-		t.Fatalf("write config: %v", err)
-	}
-	return configPath, record.Repo
-}
-
-func openMigratedCLIWebhookCoordinator(t *testing.T, dbPath string) *storage.SQLiteCoordinator {
-	t.Helper()
-
-	coordinator, err := storage.OpenSQLiteCoordinator(context.Background(), dbPath, storage.SQLiteCoordinatorOptions{BackupDir: filepath.Join(filepath.Dir(dbPath), "backups")})
-	if err != nil {
-		t.Fatalf("OpenSQLiteCoordinator() error = %v", err)
-	}
-	if _, err := coordinator.MigrationRunner().RunPending(context.Background()); err != nil {
-		t.Fatalf("MigrationRunner.RunPending() error = %v", err)
-	}
-	return coordinator
 }
